+++ CHANGELOG.md
@@ -2,9 +2,10 @@
 
 ## v3.1.0 [WIP]
 
-**Bug fix**
+**Bug fixes**
 
 -   The `Cover` component has the same image source when loading different documents
+-   The sidebar doesn't fit in its container on Safari
 
 **Breaking change**